requirements.txt
asgiref==3.8.1
attrs==24.2.0
black==24.8.0
click==8.1.7
colorama==0.4.6
dj-database-url==2.2.0
Django==5.0.7
django-filter==24.3
djangorestframework==3.15.2
drf-spectacular==0.27.2
flake8==7.1.0
gunicorn==22.0.0
h11==0.14.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.2.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycodestyle==2.12.0
pyflakes==3.2.0
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
uvicorn==0.30.5
